Command and queries for Reply domain model added.

+++ Src/Application/Themes/Commands/DeleteTheme/DeleteThemeCommandHandler.cs
@@ -1,4 +1,6 @@
-﻿using Application.Common.Interfaces;
+﻿using Application.Common.Exceptions;
+using Application.Common.Interfaces;
+using Domain.Entities;
 using MediatR;
 using System.Threading;
 using System.Threading.Tasks;
@@ -17,6 +19,9 @@ namespace Application.Themes.Commands.DeleteTheme
         public async Task<Unit> Handle(DeleteThemeCommand request, CancellationToken cancellationToken)
         {
             var theme = await _context.Themes.FindAsync(request.ThemeId);
+
+            if (theme == null)
+                throw new NotFoundException(nameof(Theme), request.ThemeId);
 
             _context.Themes.Remove(theme);
